requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Werkzeug==2.2.3
python-dotenv==0.19.0
email-validator==1.1.3
PyJWT==2.3.0 